Extend live-migration-force-complete to use postcopy if available
Since QEMU 2.5 and Libvirt 1.3.3 it is possible to use postcopy
live migration to ensure an upper limit in the amount of memory
to be transferred duing the live migration process. This option
enables to live migrate instances with memory-intensive workloads
by driving the migration process by the destination host, i.e.,
after the switch to postcopy mode, the destination VM is the active
one and therefore any dirty pages are directly generated at the
destination hosts. This patch ensures that force-complete will use
postcopy instead of pausing the VM during the migration is postcopy
is available and enabled.

Partially-implements: blueprint auto-live-migration-completion

Change-Id: Ic27f10682024ad6fee7d433f47e84aeeaa55120b
Signed-off-by: Luis Tomas <luis5tb@gmail.com>
Co-Authored-By: Pawel Koniszewski <pawel.koniszewski@intel.com>