Create _validate_requested_port_ids
This method is renamed from _process_requested_networks

That is because mostly what that method now does is to fetch each of the
requested ports for neutron, and checks if they are OK to attach to this
instance.

To save confusion about what is really returned, I have stopped
returning the network_ids, as that can be determined from the
ordered_networks.

Unit tests have been added that cover just this method in isolation from
the rest of the allocate_for_instance tests.

blueprint prep-for-network-aware-scheduling

Change-Id: I325c90e466fa1a6e7a9b844853b3b901c1243412