Pre-load tags when showing server details
When microversion>=2.26 is requested to show server details
we should pre-load 'tags' when getting the instance from the
database. Listing server details already does this, but show
didn't.

Failing to do this can also result in a 500 response from the
server. The scenario is you delete a server and then poll for
it to be gone using 'nova show'. In between the time that the
server is retrieved from the database and the view builder
creates the response, the instance is deleted and since
instance.tags is not set it will attempt to lazy-load the tags
which will fail with an InstanceNotFound.

We need to also fix the lazy-load issue in the instance object
but that will come in a follow-on change.

Change-Id: Iae6551028179e31699c06d06284ca4c1660be240
Closes-Bug: #1592963