Don't attempt to lazy-load tags on a deleted instance
Since we don't have a specific handler for lazy-loading
tags on an Instance object it loads generically but
that will fail with an InstanceNotFound on a deleted
instance, so just handle this in obj_load_attr like we
do for 'services'. Note we can't call the TagList object
to load the tags for us since the DB API will fail if
the instance is deleted too.

Change-Id: If3580dcc3e83f612b96a2c3ad893a5045ee6100a
Related-Bug: #1592963