Hyper-V: properly handle UNC instance paths
If the configured instances dir is an UNC path such as
\\ip\share\instances_dir, the HyperV driver fails to query the available
disk size, as it expects a path containing a drive letter.

Also, when attempting to move instance files during live migrations, it
will incorrectly try to build the remote path.

This change addresses those issues and is part of a bigger series,
attempting to fix HyperV shared storage related issues.

Change-Id: Ibdb9f7038bf5078867d64aef7fc63974ed8482af
Partial-Bug: #1565895