Don't immediately null host/node when shelving
When offloading a shelved instance, resources should be freed. However,
the ability to free resources is dependant on being able to find the
resource tracker for an instance's node. At present, the instance node
and host are nulled before attempting to update the resource tracker,
meaning the resources are never actually freed. Fix this by nullifying
these values *after* resources updates.

Change-Id: I8f91367aacca0c7c673b28b3c844c70c0d12f0a5
Related-bug: #1545675