remove support for legacy v2 generator extensions
This removes the support for legacy v2 extensions which were allowed
to be generators to have a pre / post processing phase. The modern
compute api stack never supported this construct.

Because of the remove of the 'pre' phase, the 'post_process' phase is
just renamed to 'process'.

All pre_process and generator tests are removed.

Change-Id: Ia34c1f814fb938915d74c6845dfa5135cba29d0a