libvirt: Rename Backend snapshot and image
snapshot and image didn't accurately describe what the methods did.
snapshot in particular suggested that the returned object was in some
way specific to a snapshot, whereas snapshot was simply the only user.

We also take the opportunity to rename all variables we touch in
driver to refer to disks rather than images.

Implements: bp/libvirt-imagebackend-refactor
Change-Id: I6ab8eb3b2c8a6af4c7ad6fba1f7d13716e9e0c12