Fix test_ipv6 and simplify to_global()
The tests test_to_global_with_bad_prefix and
test_to_global_with_bad_project are broken, they assert the TypeError
because they call to_global with an IPv6 address as MAC parameter and
do not check the functionality that they claim to check.

test_to_global_with_bad_project gets removed, because there is no
checking on the project_id, we simply hash whatever is there.

We also make sure that the correct error message is generated and add
a check to prevent regressing again.

While at it, simplify the code for to_global() in nova/ipv6/rfc2462.py.

Change-Id: I2a6f75beca1c37d135dcbb3a6eb0d739f2dd41e2