Refactor block_device_mapping handling during boot
It has become clear that it will be necessary to store a copy of the
block_device_mapping on the BuildRequest object so that it can be
returned via the API(os-extended-volumes). In order to prepare for doing
so the block_device_mapping needs to be validated and fully set, but not
persisted to the database. This patch refactors some methods to allow
that.

Change-Id: If2020f00e15ca1b067620f880a6f7d3a8ad5281c