Fix require thread policy for multi-NUMA computes
CPU thread policy "require" demands that instances are created on hosts
with hardware threads, e.g HyperThreading. It is possible that a compute
has several NUMA nodes, but only some of them contribute to the list of
pinned CPUs. In this case NUMATopologyFiler will exclude the compute
from candidates even if hardware threads are enabled and cores from
other NUMA nodes are available for CPU pinning.

This commit allows a compute with enabled HyperThreading to pass
NUMATopologyFilter if at least one NUMA node contributes to the list of
cores allowed for CPU pinning.

Change-Id: I3235cc6bf1a2a9e29a17acc43472e2a6bd4c26ec
Closes-Bug: 1596692