Removed enum duplication from nova.compute
There are many state constants held within nova.compute. Because these
now need to be used in versioned notifications, they were converted into
enums within nova.objects.fields. In order to prevent duplication, the
constants in nova.compute were converted over to use the enums created
for the notifications.

Also, because the comments in nova.compute.vm_states were not the best,
they were rearranged to more appropriately fit the lines they were
attempting to comment.

The object hash changes are ignored as it is only caused by the new
sentinel enum value _UNUSED which is not intended to be used.

Change-Id: I4c147a1a1145295b6ccfe9680706abe34cb35170