Add the 'min' param to IntOpts where applicable
It was pointed out [0] that any IntOpt with a logical minimum value
should have that specified by including the 'min=' parameter in the
option definition. As that patch was concerned with the help text for
options, it was felt that a separate follow-up patch would be better for
adding these parameters.

[0] https://review.openstack.org/#/c/281901/7/nova/conf/network.py@259

Blueprint centralize-config-options-newton

Change-Id: I0dbef37736f72401268f3036f77f35a4bbf2b5b7