Add functional regression test for bug 1595962
If the graphical consoles VNC and SPICE are disabled, the live-migration
of an instance will result in an ERROR state. VNC and SPICE are usually
disabled on IBM z systems platforms where graphical consoles are not
available. The serial console is then enabled and VNC + SPICE are
disabled. The logged error message will be:

    "Migration error: Your libvirt version does not support the
    VIR_DOMAIN_XML_MIGRATABLE flag or your destination node does
    not support retrieving listen addresses. In order for live
    migration to work properly you must either disable serial
    console or upgrade your libvirt version."

This functional test executes the live migration flow with enabled serial
console and disabled VNC + SPICE and reproduces bug 1595962.

Related-Bug: 1595962

Change-Id: Ib1b22ce77d8a36e55944acc1fbca042906aa41ef