Log DB exception if VIF creation fails
If there is a DBError raised during VIF creation it is nearly impossible
to debug since all exception information is lost. This logs the
exception to aid in debugging. Note that this exception should only be
raised if something is really wrong so this should not spam a normal
setup.

Change-Id: Ifed88c981cc9d564a7443744bf298677a0782cee