Add a py35 environment to tox
Infra will start running python35 jobs soon [1], targeting to replace
the current python34 jobs. This is due to the change from Ubuntu Trusty
to Xenial, where only python3.5 will be available.

So we add a py35 environment with the same definitions for skipping
not-yet-working checks as we have in place for py34 already.

[1] https://review.openstack.org/336272

Change-Id: I017478fafa31121f40c95d73e49ac4dced80c7a2