Throw exception if numa_nodes is not set to integer greater than 0
As [1] is abandoned, I used that patchset to create a new one.
This patchset is freshened to the current master branch.

This patch introduces InvalidNUMANodesNumber exception, which is
thrown when trying to boot an instance with a flavor that has
hw:numa_nodes=0 extra spec set. That means that NUMA nodes is set to 0,
which is incorrect.

[1]: https://review.openstack.org/#/c/190267

Change-Id: I6bd8f69e582c537a5fec40064638a8887a08cac4
Co-Authored-By: Karim Boumedhel <karimboumedhel@gmail.com>
Closes-Bug: #1402709