Deprecate nova-manage vm list command
The nova-manage vm command is replaced with the nova list command
in python-novaclient, and has been for a long time.

As of microversion 2.3, the fields that are output from
nova-manage vm list are all covered in the REST API for showing server
details, which can also be used via the --fields option of the nova
list command. The nova list command also allows filtering by host.

This sets the timer for deprecation and then removal in Ocata.

Change-Id: Ibce8c3deb24a16019b721d3b91640ca342ae541b
Closes-Bug: #1494841