Provide more duplicate VLAN network error info
When user tries to create a network which VLAN ID is already
used by another network, Network object will throw a DuplicateVlan
exception. However, when the exception get propagate to the API level,
it is only treated as a regular exception. This fix is to treat Duplicate
Vlan exception as a special exception (similar to DuplicateCidr) so that
users know more about what kind of duplicated setting they configured.

Change-Id: Ib2f7bc90957d4ae5647adbbda05a87195549374d
Closes-Bug: #1420112
Co-Authored-By: kahou Lei<kahou82@gmail.com>