Clean up helper methods in ResourceProvider
This change moves several of the module level methods that perform
database interaction back into the ResourceProvider class. They had
been created outside the class because of issues with the fixtures
python library when monkeypatching staticmethods and testing with
python3. Those issue have since been fixed in fixtures by
https://github.com/testing-cabal/fixtures/pull/30

Note that the methods that are associated with _update, _add and
_set_inventory have _not_ been moved because the changes involved
would have been more than a simple copy and paste.

Change-Id: I14395ba84cea9effe81e42f7615101ad9fab80e3
Partially-Implements: blueprint generic-resource-pools