Fix up legacy resource fields in simple-tenant-usage
This follows the previous patch and makes simple-tenant-usage use
the bundled flavor for basic instance resource usage like everything
else. This is an important step to being able to remove those fields
at some point.

Also, this often-neglected API's tests were still mocking at the
DB layer, so this converts those to be object-friendly as well.

Change-Id: Icf7d87021ea59e0197e4b15595c58550093ecdf9