Make hyperv driver use flavor fields instead of legacy ones
This makes the hyperv driver use instance.flavor.root_gb instead
of the legacy instance.root_gb field (and friends).

Also note I believe I found a bug in the
test_cache_rescue_image_bigger_than_flavor test, which was not
properly returning an image size in bytes. This was passing only
because the fake_instance object had a root_gb of 0, which made
the unit disparity moot.

Change-Id: I03207a267f1ffb21ea464165c1910f66962caab6