Validate pci_passthrough_whitelist when starting n-cpu
Loading up CONF.pci_passthrough_whitelist in the Whitelist
object performs a bunch of validation and can fail in several
different ways (invalid json, invalid values, invalid combinations
of keys, devices not found, etc). This happens today when
creating the PciDevTracker in the ResourceTracker when updating
available resources. If the configuration is bad, it kills the
periodic task to update available resources on the compute node.

We should just load up the pci_passthrough_whitelist (if set)
when starting the nova-compute service so we can fail fast and
kill the service on any misconfiguration rather than run with
a broken service.

Change-Id: If50fb837b490042bb5ef20e9ad843b28f871a44e
Closes-Bug: #1603034