Allow linear packing of cores
Given the following single-socket, four-core, HT-enabled CPU topology:

   +---+---+ +---+---+ +---+---+ +---+---+
   | x | x | | x |   | | x |   | |   |   |
   +---+---+ +---+---+ +---+---+ +---+---+
     1   4     2   5     3   6     4   7

Attempting to boot an instance with four cores and no explicit
'cpu_thread_policy' should be successful, with cores 5,6,4,7 used.
However, the current implementation of this implicit policy attempts to
fit the same number of instance cores onto each host CPU. For example,
a four core instance would result in either a 2*2 layout (two instance
cores on each of two host CPUs), or a 1*4 layout (one instance core on
each of four host CPUs). This may be correct behavior *where possible*,
but if this is not possible then any and all cores should be used.

Resolve this issue by adding a fallthrough case, whereby if the
standard fitting policy fails, a linear assignment is used to properly
fit the instance cores.

Change-Id: I73f7f771b7514060f1f74066e3dea1da8fe74c21
Closes-Bug: #1578155
mitaka-backport-potential