Fix exception due to BDM race in get_available_resource()
If we run the resource tracker periodic at the right time, we
may try to collect BDM info from a newly-created instance before
we have any BDM records for it. This patch excludes instances
that have no reported BDMs to avoid choking there. This also
adds a test which simulates an instance that is partially in
the database, but is not fully created.

Closes-Bug: #1602057
Change-Id: I12c9c1ae6ca27727e8742060647dbe7017cded08
(cherry picked from commit 66246c4c9b6f766f40ee922c38c46f35bb02ae70)