Fix database poison warnings, part 9
The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    unit.api.openstack.compute.test_keypairs.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I4284dd7ecb9c60834dba662838f67d9defc798e2
Related-Bug: #1568414