Fix database poison warnings, part 11
The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes the one warning from:

    unit.virt.libvirt.test_imagecache.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I057852125249d41d0b645a68013fef62249a2af7
Related-Bug: #1568414