Free new pci_devices on revert-resize
On confirm-resize the drop_move_claim is called
with prefix="old_" which means it should drop the
old flavor allocations.
On revert-resize the drop_move_claim is called
with prefix="new_" which means it should drop
the new flavor allocations.

This patch update that the drop_move_claim will
drop pci devices according to the prefix

Co-Authored-By: Ludovic Beliveau <ludovic.beliveau@windriver.com>

Closes-Bug: #1607219

Change-Id: I4601f89d8ab23db19b7515d0ad9779638ec66b49