Fix database poison warnings, part 14
The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes the warnings from:

    unit.virt.xenapi.test_xenapi.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I41537e6b062b169cc6934d525560eb96ca58c55b
Related-Bug: #1568414