rt: don't log pci_devices twice when updating resources
By default we update available resources on the compute every
60 seconds. The _report_hypervisor_resource_view method is
logging pci devices twice. On a compute with hundreds of pci
devices this blows up the logs in a short amount of time.

This change removes at least the duplicate pci device logging
but we might want to even consider just not logging these in
each iteration. I've opted to at least push the simple fix for
now.

Change-Id: Id05bfef44b1108dec286486d42516649dd0683e9
Closes-Bug: #1606718