Fix drop_move_claim() on revert resize
On revert_resize(), the migration context was dropped hence setting it
to None.  The problem is that the migration context is needed when
executing finish_revert_resize() so that the neutron port can be updated
with the right binding profile.

This patch moves the instance.drop_migration_context() call out from the
resource tracker so that the migration context can be dropped at the
appropriate times: after RT.drop_move_claim() is called in the compute
manager's confirm_resize() call and after the virt driver's
finish_revert_resize() in the compute manager's finish_revert_resize()
call.

Change-Id: I1b7b2573de4380576dd8b801ed59d8955b0ab72a
Partial-bug: #1512880