no instance info cache update if instance deleted
Avoid instance info cache update if the instance is deleted,
as the deleted instance's info cache is already deleted in
db layer and it will report InstanceInfoCacheNotFound
exception and lead to a few exception logs in compute.

Please note this is different to regular info cache and
lead to InstanceInfoCacheNotFound exception, it's in the
case when you first create the info cache after instance
delete and InstanceInfo is also deleted.

Change-Id: I860e9e7c7ef458722135a21c6c5745f5519c56c4
Closes-Bug: 1618822