Straddle python-neutronclient 5.0 for testing
python-neutronclient 5.0 has deprecated the usage of
tenant_id/tenant_name in place of project_id/project_name
for keystone v3 usage. It does this using debtcollector
decorators on the methods with those kwargs. This is fine
at runtime but doesn't work with mox since mox stubs the
methods off the Client class, but doesn't pick up and stub
any decorators on those methods. So when our unit tests
pass tenant_id kwargs they fail on the moxed_client.

To make the tests work for both neutronclient>5 and <5 we
simply pass the project_id as an arg rather than kwarg
to the show_quota method.

Change-Id: I86bbac3e24bef27396297dfc79eb24c61dbcb6ac
Closes-Bug: #1608258