Changed exception catching order
On [1] InstanceIsLocked is catched. However on [2] InstanceInvalidState
is catched, which is an ancestor of InstanceIsLocked.
For this reason, catching a specific exception is meaningless after
catching the base type (as it will catch sub-types too).

So I changed the order of the catch clauses, which now
catches InstanceIsLocked before InstanceInvalidState, which means
the more specific exception is catched before the more general
exception.

[1]: https://github.com/openstack/nova/blob/master/nova/api/openstack/compute/servers.py#L1111
[2]: https://github.com/openstack/nova/blob/master/nova/api/openstack/compute/servers.py#L1108

Change-Id: I4c520d485736197c278291c126784f7a8d693756