libvirt: Create console.log consistently
Change Ia84579b8 originally removed creation of console.log by Nova,
but this was found to break Quobyte CI (bug 1597644), and the change
was reverted. _create_image continues to be the wrong place for this
code, however, as it is not related to the creation of instance disks,
only the initialisation of a new instance directory. Note that this
does not correlate 1-1 with callers of _create_image: it is not
required by rescue(), as console.log will already exist.

The old code used a different method to create console.log for live
migration. We clean this up to have a single helper function.

The old code chowned console.log in the rescue case, which was never
required as the log is chowned immediately before use in
get_console_output(). It is not required in the Quobyte CI case
either, as libvirt has been configured with dynamic_ownership=0, so it
has not been chowned since creation. As
_chown_console_log_for_instance() has only 1 remaining caller, it is
inlined in get_console_output(), which also removes some duplication.

Note that I am not personally convinced that the Quobyte CI usage is
valid or correct, but I'm happy to not break it in the way we
currently know about while that determination is made. In the
meantime, I still need to move the console.log manipulation code out
of the way to refactor _create_image().

Change-Id: Iecf543b3cbf83df6c4a6ad17ce805979198556c8