rt: Refactor resize_claim unit test
I don't care for the way the migration and resize unit tests were done
in the test_tracker.py file. It isn't clear what is being asserted in
the unit tests because of layers of nested helper functions. This patch
is the start of a series that refactors these tests to be more readable.
First up was the unit test for resize_claim on the same host.

Change-Id: I370d32e1019208adacdb17ae75d87a8711e4d8c4