rt: Remove duplicate unit test for missing mig ctx
Removes a unit test that purported to test that "a resource audit
occurred" even if an instance's migration_context attribute was missing,
as would be the case "on older computes". The unit test wasn't actually
testing this though. All the unit test was testing was essentially the
lazy-load behaviour of the nova.objects.Instance.migration_context
attribute which is already tested in nova/tests/unit/objects/instance.py
here:

https://github.com/openstack/nova/blob/\
8d007b4c30ac8d32da181fab06563c5442e0f3a9\
/nova/tests/unit/objects/test_instance.py#L1299-L1325

Change-Id: I168659fbc5cb3fc064013695210bf0b223ac02a8