Fix corrupt "host_aggregates_map" in host_manager
A host can be in multiple host-aggregates at the same time. When a
host gets removed from an aggregate in thread A and this aggregate
gets deleted in thread B, there can be a race-condition where the
mapping data in the host_manager can get out of sync for a moment.

This change simulates this condition in a unit test and fixes the bug
by iterating over the mapping itself instead of the out-of-date list
"aggregates.hosts".

Closes-Bug: 1605804
Change-Id: I59861f03f0c681f7118782fb017af377e07552aa