Replace mox with mock in test_cinder
Replace mox with mock in test_cinder to complete one task
in the old blueprint. This patch also intends to partially
reduce possible race conditions as shown in the following
log file:

http://logs.openstack.org/58/335358/12/check/gate-nova-python34-db/404641d/console.html#_2016-08-08_19_47_57_286193

Further modifications are contained by the follow up patch
in the chain to add fixes to test_compute as well.

Implements: blueprint mox-to-mock-conversion
Change-Id: I5f27c65c405131768bb8d30b84c5cddb6ace1d27