rt: New unit test for rebuild_claim()
There was a unit test in test_resource_tracker.py called
test_existing_migration() in the _MoveClaimTest class. Ostensibly, this
unit test was supposed to be testing when the _move_claim() method of
the resource tracker was being called with an explicit migration object
that had been passed in from the compute manager. Unfortunately, either
due to the passing of time and corrosion of the unit tests in
test_resource_tracker.py or simply an oversight by the original test
author, the resource tracker's _move_claim() method is never called
directly and the only method that calls _move_claim() with an explicit
migration parameter is the rebuild_claim() method. This method wasn't
being tested at all, in either test_resource_tracker.py or
test_tracker.py.

This patch adds that missing unit test of rebuild_claim() by extracting
the original test_existing_migration() unit test from
test_resource_tracker.py that wasn't really testing anything that could
be called directly from the compute manager and changing it to conform
to the test fixtures in test_tracker.py.

Change-Id: I3024edc7b56e652637c31b209160c72347c1f50d