rt: Create multiple resize claim unit test
Creates a new unit test in test_tracker.TestResize for testing the
resource accounting when two resize claims hit a destination node one
after the other. There was a single remaining unit test in
test_resource_tracker.py called test_additive_claims() that was trying
to test this situation but doing it in a way that wasn't particularly
readable or understandable.

Change-Id: Ifb8e4de975723a04035b262d182409a9078f46e8