Add separate create/delete policies to attach_interface
In the v2 API there were separate policy checks for the attach and
detach interface actions. This allowed deployers to allow one and not
the other. The v2.1 API policy check did not have this split so both had
to be enabled/disabled.

This patch adds additional checks to allow deployers more granular
control.

Change-Id: Icf1f0dd12920a2c6126e52a548f3fa4636b431d6
Closes-Bug: 1610069