Add cell_v2 verify_instance command
This adds a verify_instance command to nova-manage, which simply fetches
out the instance and cell mapping for a given instance uuid and prints
the result. This is useful to sanity-check a deployment after an upgrade,
and will be used by grenade to ensure that the mappings were created
after running the migration command.

Change-Id: I4aaff99d59e8b5f530a30f5aea0be195b246d9f7