Make simple_cell_setup idempotent
This makes it so we can run the simple_cell_setup command multiple
times without failure. This would be important for things like grenade,
as well as deployment tools that need to just run it and might not
know if they're upgrading or deploying a new setup.

This rearranges the order in the simple_cell_setup command to do the
cell0 mapping first, to avoid running the host mapping other than the
very first time. Before, if the user was willing to tolerate the error,
they could run this command over and over and continue to map hosts
into cells. Since this change is encouraging the ability to run this
implicitly in multiple places, we want to defeat that behavor so that
it's not abused.

Change-Id: I559f9c87e89926414b368cac9442dec4eadcb89b