os-vif: do not set Route.interface if None
The Route.interface field in os-vif is currently
non-nullable, which means it must be something other
than None if set.

There is no code in nova that sets the interface field
on the nova.network.model.Route object, so it defaults
to None and is always set to None on the os-vif Route
object, which fails with:

"ValueError: Fieldinterface' cannot be None"

We don't hit this in the gate because we don't test with
subnets that have host_routes set.

Change-Id: Ideadf1fc009b3ee496acb4696226a3cc9722fb29
Closes-Bug: #1612812