Fix neutron security group tests for 5.1.0 neutronclient
python-neutronclient 5.1.0 has a different API call for
find_resourceid_by_name_or_id now:

84c42160e9eac654e734a19032237fdf1a8cb058

That breaks the neutron security group driver unit tests
which are relying on a mocked out version of the neutronclient
which itself is a bad idea but not completely re-written here

This blocks getting to use neutronclient 5.1.0 in
upper-constraints so this change just fixes the immediate
breakage.

Also removes test_get_raises_no_unique_match_error which
duplicates the coverage from test_associate_duplicate_names
for the 409 scenario.

Change-Id: Iee340361bc656a0bd3a246e02de41a8f4f89bebf
Closes-Bug: #1612494